Organize shared workspace icons
